feat(tabs): move shift-tabs into tabs category
