+++ content/docs/tabs/shift-tabs.mdx
@@ -5,24 +5,24 @@ labels: ["requires interaction", "hover"]
 author: morphhyy
 ---
 
-<ComponentPreview name="container-shift-tabs--primary" />
+<ComponentPreview name="tabs-shift-tabs--primary" />
 
 ## Installation
 
 <Steps>
 <Step>Run the following command</Step>
 
-It will create a new file called `shift-tabs.tsx` inside the `components/animata/container` directory.
+It will create a new file called `shift-tabs.tsx` inside the `components/animata/tabs` directory.
 
 ```bash
-mkdir -p components/animata/container && touch components/animata/container/shift-tabs.tsx
+mkdir -p components/animata/tabs && touch components/animata/tabs/shift-tabs.tsx
 ```
 
 <Step>Paste the code</Step>
 
 Open the newly created file and paste the following code:
 
-```jsx file=<rootDir>/animata/container/shift-tabs.tsx
+```jsx file=<rootDir>/animata/tabs/shift-tabs.tsx
 
 ```
